--- AgenciaDeViajes/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\Dz\\AppData\\Roaming\\NetBeans\\11.0\\build.properties
