--- AgenciaDeViajes/nbproject/project.properties
@@ -1,75 +1,0 @@
-#Mon Nov 01 14:04:50 COT 2021
-excludes=
-javac.test.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}:\
-    ${libs.junit_5.classpath}:\
-    ${libs.junit_4.classpath}:\
-    ${libs.hamcrest.classpath}
-run.classpath=${javac.classpath}\:${build.classes.dir}
-javac.processorpath=${javac.classpath}
-dist.javadoc.dir=${dist.dir}/javadoc
-test.src.dir=test
-run.modulepath=${javac.modulepath}
-annotation.processing.enabled=true
-build.sysclasspath=ignore
-debug.modulepath=${run.modulepath}
-javac.compilerargs=
-javadoc.noindex=false
-javadoc.private=false
-javadoc.author=false
-main.class=agenciadeviajes.ClienteMain
-junit.selected.version=4
-source.encoding=UTF-8
-javac.source=12
-includes=**
-javadoc.use=true
-jar.compress=false
-javadoc.nonavbar=false
-annotation.processing.enabled.in.editor=false
-javadoc.notree=false
-annotation.processing.processors.list=
-javac.deprecation=false
-jlink.launcher=true
-javadoc.additionalparam=
-jlink.launcher.name=AgenciaDeViajes
-build.generated.sources.dir=${build.dir}/generated-sources
-javadoc.splitindex=true
-javac.processormodulepath=
-run.jvmargs=
-jlink.additionalmodules=
-javadoc.encoding=${source.encoding}
-javac.classpath=${libs.junit_5.classpath}\:${libs.junit_4.classpath}
-mkdist.disabled=false
-run.test.modulepath=${javac.test.modulepath}
-build.classes.excludes=**/*.java,**/*.form
-dist.jlink.dir=${dist.dir}/jlink
-dist.jar=${dist.dir}/AgenciaDeViajes.jar
-build.classes.dir=${build.dir}/classes
-debug.test.modulepath=${run.test.modulepath}
-build.test.classes.dir=${build.dir}/test/classes
-javadoc.windowtitle=
-build.test.results.dir=${build.dir}/test/results
-annotation.processing.processor.options=
-dist.dir=dist
-build.dir=build
-dist.archive.excludes=
-annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-build.generated.dir=${build.dir}/generated
-javadoc.version=false
-javac.test.modulepath=${javac.modulepath}
-debug.test.classpath=${run.test.classpath}
-javac.external.vm=true
-javac.target=12
-platform.active=default_platform
-manifest.file=manifest.mf
-javadoc.html5=false
-javac.test.processorpath=${javac.test.classpath}
-meta.inf.dir=${src.dir}/META-INF
-run.test.classpath=${javac.test.classpath}\:${build.test.classes.dir}
-dist.jlink.output=${dist.jlink.dir}/AgenciaDeViajes
-annotation.processing.run.all.processors=true
-javac.modulepath=
-src.dir=src
-jlink.additionalparam=
-debug.classpath=${run.classpath}
