+++ AgenciaDeViajesMaven/README.md
@@ -1,0 +1,13 @@
+# Practica 1: Ejercicio de Polimorfismo 
+## Realizado en Java - Maven
+## Integrantes: Anderson Diaz Montenegro
+
+
+## Para construir el .jar: 
+```
+mvn package 
+```
+## Para ejecutar el test: 
+```
+mvn test
+```
