+++ AgenciaDeViajesMaven/src/main/java/LogicaNegocio/ViajeFamiliar.java
@@ -1,0 +1,41 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package LogicaNegocio;
+
+import java.util.Date;
+
+/**
+ *
+ * @author Dz
+ */
+public class ViajeFamiliar extends Viaje {
+
+    private int familia;
+
+    //Constructor getters and setters
+    public ViajeFamiliar(String origen, String destino, int costo, Date fechaSalida, Date fechaLlegada, int familia) {
+        super(origen, destino, costo, fechaSalida, fechaLlegada);
+        this.familia = familia;
+    }
+
+    public int getFamilia() {
+        return familia;
+    }
+
+    public void setFamilia(int familia) {
+        this.familia = familia;
+    }
+
+    @Override
+    public String descripcion() {
+        return "Viaje para disfrutar con toda tu familia";
+    }
+
+    @Override
+    public String cualquierMetodo2() {
+        return "Método implementado en la clase hija viaje familiar";
+    }
+
+}
