--- README.txt
@@ -1,3 +1,0 @@
-Anderson Diaz Montenegro
-Primera Practica - Polimorfismo
-1/11/2021
